% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataRegistry.R
\name{add_table}
\alias{add_table}
\title{Add a Table's Metadata to a Registry}
\usage{
add_table(registry, table_metadata)
}
\arguments{
\item{registry}{The registry object to which metadata will be added.}

\item{table_metadata}{A data.table object created by \code{table_info()}.}
}
\value{
The updated registry object.
}
\description{
A generic function to add new table metadata to a registry object.
}
