% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{attendance}
\alias{attendance}
\title{Sample Attendance Data}
\format{
A data.table with 5 rows and 2 variables:
\describe{
\item{id}{Character. Unique identifier for each student.}
\item{attendance_days}{Integer. Number of days the student was present.}
}
}
\source{
This data is manually created for demonstration purposes.
}
\usage{
attendance
}
\description{
A dataset containing student IDs and their corresponding attendance records.
}
\examples{
data(attendance)
head(attendance)

}
\keyword{datasets}
