% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataRegistry.R
\name{create_metadata_registry}
\alias{create_metadata_registry}
\title{Create a Metadata Registry}
\usage{
create_metadata_registry()
}
\value{
An empty data.table with the class "MetadataRegistry".
}
\description{
Initializes an empty data.table with a custom class "MetadataRegistry" to
store and manage metadata definitions.
}
