% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customers}
\alias{customers}
\title{Sample Customer Data}
\format{
A data.table with 5 variables:
\describe{
\item{customer_id}{A unique identifier for each customer.}
\item{age}{The age of the customer in years.}
\item{gender}{The gender of the customer.}
\item{income}{The income level of the customer.}
\item{region}{The geographical region where the customer resides.}
}
}
\source{
Generated for package examples.
}
\usage{
customers
}
\description{
A sample dataset containing demographic information for customers included with
the DBmaps package.
}
\keyword{datasets}
