% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_plan.R
\name{execute_join_plan}
\alias{execute_join_plan}
\title{Execute a Join Plan}
\usage{
execute_join_plan(join_plan, data_list)
}
\arguments{
\item{join_plan}{A data.table created by \code{create_join_plan()}.}

\item{data_list}{A named list of the source data.tables.}
}
\value{
A final, merged data.table.
}
\description{
Takes a plan generated by \code{create_join_plan()} and executes it sequentially
to produce a final, merged data.table.
}
