% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_all_tables.R
\name{merge_all_tables}
\alias{merge_all_tables}
\title{Merge Three Data Tables}
\usage{
merge_all_tables(
  x,
  y,
  z,
  by_col = "id",
  merge_type = "inner",
  final_merge_type = "inner"
)
}
\arguments{
\item{x}{First data table (e.g., students).}

\item{y}{Second data table (e.g., scores).}

\item{z}{Third data table (e.g., attendance).}

\item{by_col}{The column name (as a string) on which the merge should be
performed. Default is \code{"id"}.}

\item{merge_type}{Type of merge to perform between \code{x} and \code{y}. Options:
\itemize{
\item \code{"inner"}: Matches only common rows.
\item \code{"left"}: All rows from \code{x}, with matches from \code{y}.
\item \code{"right"}: All rows from \code{y}, with matches from \code{x}.
\item \code{"full"}: All rows from both.
}}

\item{final_merge_type}{Type of merge to perform between \code{(x + y)} and \code{z}.
Same options as \code{merge_type}.}
}
\value{
A merged \code{data.table} object containing all three tables.
}
\description{
This function first merges two data tables (\code{x} and \code{y}) using the specified
join type. It then merges the resulting table with a third data table (\code{z}).
}
\examples{
merge_all_tables(students, scores, attendance, merge_type = "inner",
      final_merge_type = "full")
merge_all_tables(students, scores, attendance, merge_type = "left",
      final_merge_type = "right")

}
