% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_join_plan.R
\name{plot_join_plan}
\alias{plot_join_plan}
\title{Plot a Join Plan as a Flowchart}
\usage{
plot_join_plan(join_plan)
}
\arguments{
\item{join_plan}{A \code{data.table} created by \code{create_join_plan()}.}
}
\value{
A \code{DiagrammeR} graph object that can be printed to the RStudio
Viewer pane.
}
\description{
Takes a plan generated by \code{create_join_plan()} and creates a flowchart
visualizing the sequence of aggregations and merges.
}
