% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{products}
\alias{products}
\title{Sample Product Data}
\format{
A data.table with 3 variables:
\describe{
\item{product_id}{A unique identifier for each product.}
\item{category}{The category to which the product belongs.}
\item{original_price}{The original price of the product.}
}
}
\source{
Generated for package examples.
}
\usage{
products
}
\description{
A sample dataset containing product information included with the DBmaps package.
}
\keyword{datasets}
