% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{scores}
\alias{scores}
\title{Sample Student Scores Data}
\format{
A data.table with 3 rows and 3 variables:
\describe{
\item{id}{Character. Student ID}
\item{homework}{Numeric. Homework scores}
\item{quiz}{Numeric. Quiz scores}
}
}
\source{
This data is manually created for demonstration purposes.
}
\usage{
scores
}
\description{
The \code{scores} dataset contains students' homework and quiz scores.
}
\examples{
data(scores)
head(scores)

}
\keyword{datasets}
