% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{students}
\alias{students}
\title{Sample Student Data}
\format{
A data.table with 4 rows and 2 variables:
\describe{
\item{id}{Character. Student ID}
\item{Birthdate}{Date. Date of birth}
}
}
\usage{
students
}
\description{
The \code{students} dataset contains information about students and their birthdates.
}
\examples{
data(students)
head(students)
}
\keyword{datasets}
