% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transactions}
\alias{transactions}
\title{Sample Transaction Data}
\format{
A data.table with 5 variables:
\describe{
\item{customer_id}{Identifier for the customer making the transaction.}
\item{product_id}{Identifier for the product being purchased.}
\item{time}{The timestamp of the transaction (POSIXct format).}
\item{quantity}{The number of units of the product purchased.}
\item{price}{The price per unit at the time of transaction.}
}
}
\source{
Generated for package examples.
}
\usage{
transactions
}
\description{
A sample dataset of transaction events, linking customers and products. This
is a typical "fact" table in a relational schema.
}
\keyword{datasets}
