% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{views}
\alias{views}
\title{Sample Product View Data}
\format{
A data.table with 3 variables:
\describe{
\item{customer_id}{Identifier for the customer viewing the product.}
\item{product_id}{Identifier for the product being viewed.}
\item{time}{The timestamp of the view event (POSIXct format).}
}
}
\source{
Generated for package examples.
}
\usage{
views
}
\description{
A sample dataset of product view events, linking customers and products. This
is a smaller, sampled version of a potentially very large event log.
}
\keyword{datasets}
